+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/TalkCache.java
@@ -5,15 +5,22 @@ import net.minecraft.world.entity.Entity;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
+import java.util.function.Supplier;
 
 public class TalkCache {
 
     private static final long TIMEOUT = 250L;
 
     private final Map<UUID, Talk> cache;
+    private Supplier<Long> timestampSupplier;
 
     public TalkCache() {
         this.cache = new HashMap<>();
+        this.timestampSupplier = System::currentTimeMillis;
+    }
+
+    public void setTimestampSupplier(Supplier<Long> timestampSupplier) {
+        this.timestampSupplier = timestampSupplier;
     }
 
     public void updateTalking(UUID entity, boolean whispering) {
@@ -32,27 +39,31 @@ public class TalkCache {
         if (entity.equals(ClientManager.getPlayerStateManager().getOwnID())) {
             ClientVoicechat client = ClientManager.getClient();
             if (client != null && client.getMicThread() != null) {
-                return client.getMicThread().isTalking();
+                if (client.getMicThread().isTalking()) {
+                    return true;
+                }
             }
         }
 
         Talk lastTalk = cache.getOrDefault(entity, new Talk(0L, false));
-        return System.currentTimeMillis() - lastTalk.timestamp < TIMEOUT;
+        return timestampSupplier.get() - lastTalk.timestamp < TIMEOUT;
     }
 
     public boolean isWhispering(UUID entity) {
         if (entity.equals(ClientManager.getPlayerStateManager().getOwnID())) {
             ClientVoicechat client = ClientManager.getClient();
             if (client != null && client.getMicThread() != null) {
-                return client.getMicThread().isWhispering();
+                if (client.getMicThread().isWhispering()) {
+                    return true;
+                }
             }
         }
 
         Talk lastTalk = cache.getOrDefault(entity, new Talk(0L, false));
-        return lastTalk.whispering && System.currentTimeMillis() - lastTalk.timestamp < TIMEOUT;
+        return lastTalk.whispering && timestampSupplier.get() - lastTalk.timestamp < TIMEOUT;
     }
 
-    private static class Talk {
+    private class Talk {
         private final long timestamp;
         private final boolean whispering;
 
@@ -62,7 +73,7 @@ public class TalkCache {
         }
 
         public Talk(boolean whispering) {
-            this.timestamp = System.currentTimeMillis();
+            this.timestamp = timestampSupplier.get();
             this.whispering = whispering;
         }
     }